tt_stl: add scope guard utility
Move generic ScopeGuard implementation into tt_stl and re-export from ttnn.
