+++ tt_stl/CMakeLists.txt
@@ -17,6 +17,7 @@ target_sources(
             tt_stl/aligned_allocator.hpp
             tt_stl/cleanup.hpp
             tt_stl/concepts.hpp
+            tt_stl/guard.hpp
             tt_stl/indestructible.hpp
             tt_stl/optional_reference.hpp
             tt_stl/overloaded.hpp
